risk_service/requirements.txt
grpcio
grpcio-tools
sqlalchemy
psycopg2-binary